Add files via upload
Problem Definition:
It is required to simulate a limited number of devices connected to a router’s Wi-Fi
using Java threading and semaphore. Routers can be designed to limit the number of
open connections. For example, a Router may wish to have only N connections at any
point in time. As soon as N connections are made, the Router will not accept other
incoming connections until an existing connection is released. Explain how
semaphores can be used by a Router to limit the number of concurrent connections.